First attempt at smaller vision cones
The smaller vision cones don't work fully quite yet - I need to adjust the position of the cone when it is facing downward so that the cone isn't detached from the hall monitor. But this is a start.

I also adjusted the collision boxes of all the sprites so that they match the images.