Remove hardcoding from stationary monitors and bells
Lots of updates to the code in this commit. By "remove hardcoding", I mean removing the need to specify coordinates of the vision cones or hall monitors when using the custom scripts. I tried many things to remove hardcoding from the update scripts of the stationary hall monitors and vision cones, and I ran into performance issues (especially in the final level, which has a lot of actors in the scene) and GB Studio/GBVM bugs. However, I think the resulting code is still a good compromise of usability/maintainability and performance.

Next step: Do the same for the patrolling hall monitors and vision cones and the janitor carts.