Standardize janitor cart code
Changed all the janitor cart actors to use the same custom script. This means they will all behave the same way, and it reduces a lot of code in the project.

Note that I'd like to use collisions for moving the janitor carts, but due to a bug with the start menu, the player can get stuck inside a janitor cart. If this happens, the janitor cart needs to have collisions turned off so that the player can exit the janitor cart.